Create a package lookup and move Definition into the package.
