+++ modules/lang-painless/src/main/java/org/elasticsearch/painless/node/AExpression.java
@@ -20,8 +20,8 @@
 package org.elasticsearch.painless.node;
 
 import org.elasticsearch.painless.AnalyzerCaster;
-import org.elasticsearch.painless.Definition;
-import org.elasticsearch.painless.Definition.Cast;
+import org.elasticsearch.painless.lookup.Definition;
+import org.elasticsearch.painless.lookup.Definition.Cast;
 import org.elasticsearch.painless.Locals;
 import org.elasticsearch.painless.Location;
