+++ modules/lang-painless/src/main/java/org/elasticsearch/painless/node/EListInit.java
@@ -19,9 +19,9 @@
 
 package org.elasticsearch.painless.node;
 
-import org.elasticsearch.painless.Definition.Method;
-import org.elasticsearch.painless.Definition.MethodKey;
-import org.elasticsearch.painless.Definition.def;
+import org.elasticsearch.painless.lookup.Definition.Method;
+import org.elasticsearch.painless.lookup.Definition.MethodKey;
+import org.elasticsearch.painless.lookup.Definition.def;
 import org.elasticsearch.painless.Globals;
 import org.elasticsearch.painless.Locals;
 import org.elasticsearch.painless.Location;
